+++ src/blocks/Table/models.ts
@@ -1,5 +1,0 @@
-export const TableMicrodataValues = {
-    TableType: 'https://schema.org/Table',
-    AboutProp: 'about',
-    accessModeProp: 'accessMode',
-} as const;